split up lattice and utils definitions
